requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cycler==0.12.1
Django==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-crispy-forms==2.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
flower==2.0.1
fonttools==4.51.0
greenlet==3.0.3
humanize==4.9.0
idna==3.7
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.8.4
mysqlclient==2.2.4
Naked==0.1.32
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.22.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycryptodome==3.20.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
scipy==1.13.0
shellescape==3.8.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
statsmodels==0.14.2
tenacity==8.3.0
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
